+++ liana/method/sp/__init__.py
@@ -1,3 +1,3 @@
-from ._lr_bivar import lr_bivar
-from ._SpatialBivariate import bivar
-from ._misty_constructs import genericMistyData, lrMistyData, MistyData
+from liana.method.sp._bivariate._lr_bivar import lr_bivar
+from liana.method.sp._bivariate._SpatialBivariate import bivar
+from liana.method.sp._misty._misty_constructs import genericMistyData, lrMistyData, MistyData
