+++ liana/method/sp/_bivariate/_spatial_pipe.py
