+++ liana/tests/test_bivar.py
@@ -2,7 +2,7 @@ import numpy as np
 from itertools import product
 
 from liana.testing._sample_anndata import generate_toy_mdata
-from liana.method.sp._SpatialBivariate import bivar
+from liana.method.sp._bivariate._SpatialBivariate import bivar
 
 mdata = generate_toy_mdata()
 interactions = list(product(mdata.mod['adata_x'].var.index,
