+++ liana/tests/test_spatial_utils.py
@@ -1,8 +1,8 @@
 import numpy as np
 from  scipy.sparse import csr_matrix
 
-from liana.method.sp._spatial_pipe import _global_r
-from liana.method.sp._bivariate_funs import LocalFunction
+from liana.method.sp._bivariate._spatial_pipe import _global_r
+from liana.method.sp._bivariate._bivariate_funs import LocalFunction
 from liana.utils.spatial_neighbors import spatial_neighbors
 
 from liana.testing._sample_anndata import generate_toy_spatial
